+++ demo/App.js
@@ -4,7 +4,7 @@ import DropdownMenu from '../src/DropdownMenu';
 import MenuItem from '../src/MenuItem';
 import {
   Home, Simple, IconColor, ShowUser, LinkStyle, ClickEvent, Multiple, FadeIn,
-  OtherDropDown, Separator, Position, ImageTrigger, TextTrigger, IconTrigger
+  OtherDropDown, Separator, Position, ImageTrigger, TextTrigger, IconTrigger, DisabledItem
 } from './examples/Index';
 
 class App extends React.Component {
@@ -18,6 +18,7 @@ class App extends React.Component {
           <Route path="/showuser" component={ShowUser} />
           <Route path="/linkstyle" component={LinkStyle} />
           <Route path="/clickevent" component={ClickEvent} />
+          <Route path="/disabled" component={DisabledItem} />
           <Route path="/separator" component={Separator} />
           <Route path="/position" component={Position} />
           <Route path="/imagetrigger" component={ImageTrigger} />